% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth-death.R
\name{calc_bd_lm_nll}
\alias{calc_bd_lm_nll}
\title{Calculate negative log likelihood of linear model parameterization
of birth-death process}
\usage{
calc_bd_lm_nll(w, x, y, param_map)
}
\arguments{
\item{w}{vector of parameters}

\item{x}{matrix of predictors}

\item{y}{tree assumed to come from birth-death process}

\item{param_map}{function to map w to list with parameters for calc_bd_nll()}
}
\description{
Calculate negative log likelihood of linear model parameterization
of birth-death process
}

