% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth-death.R
\name{calc_bd_nll}
\alias{calc_bd_nll}
\title{Calculate the negative log likelihood of a tree occurring according
to a birth-death process.}
\usage{
calc_bd_nll(l, m, psi, freq, phylo, survival = FALSE,
  unknown_states = FALSE, rtol = 1e-12, atol = 1e-12, cutoff = 10^12)
}
\arguments{
\item{l}{Matrix of rates at which a type i individual gives birth
to a type j individual}

\item{m}{Vector of death rates for each type}

\item{psi}{Vector of death rates for each type where sampling
occurs at time of death. The probability of sampling at death is psi / (psi + m)}

\item{freq}{Expected probability of each type at the root of the
tree}

\item{phylo}{Phylo object with tree for which likelihood will be
calculated}

\item{survival}{Boolean indicator of whether likelihood is divided
by likelihood of tree being observed}

\item{unknown_states}{Boolean indicator of whether the type of the
tree tips is unknown}

\item{rtol}{Passed through to deSolve differential equation solver}

\item{atol}{Passed through to deSolve differential equation solver}

\item{cutoff}{Time in past from most recent tip beyond which
sampling probability assumed zero}
}
\description{
Calculate the negative log likelihood of a tree occurring according
to a birth-death process.
}

