% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{calc_phylo_nll}
\alias{calc_phylo_nll}
\title{Calculate the negative log likelihood of a time scaled phylogeny}
\usage{
calc_phylo_nll(w, x, y, param_map)
}
\arguments{
\item{w}{vector of parameters}

\item{x}{matrix of predictors}

\item{y}{tree assumed to come from birth-death process}

\item{param_map}{function to map w to list with parameters for the
phylogenetic likelihood calculation}
}
\description{
Calculate the negative log likelihood of a time scaled phylogeny
}

