% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{eval_temporal_signal}
\alias{eval_temporal_signal}
\title{Evaluate signal of substitutions per unit time}
\usage{
eval_temporal_signal(phy, tipheights, show_plots = FALSE)
}
\arguments{
\item{phy}{rooted tree with branch lengths in substitutions per site}

\item{tipheights}{heights of tree tips in time}

\item{show_plots}{whether or not to show diagnostic plots for
root-to-tip regression}
}
\description{
Evaluate signal of substitutions per unit time
}

