% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth-death.R
\name{gen_param_map}
\alias{gen_param_map}
\title{Generate parameter map for linear model interface to birth-death nll}
\usage{
gen_param_map(n, ntrees, psampled = 0.01)
}
\arguments{
\item{n}{number of types in birth-death model}

\item{ntrees}{number of trees that will be fitted to birth-death model}

\item{psampled}{proportion of deaths that are assumed to be
observed in birth-death model}
}
\description{
Generate parameter map for linear model interface to birth-death nll
}

