% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{gen_param_map_phylo}
\alias{gen_param_map_phylo}
\title{Generate parameter map for phylogeny nll}
\usage{
gen_param_map_phylo(tree, tip_times)
}
\arguments{
\item{tree}{tree with branchlengths in time units}

\item{tip_times}{times of tips of tree}
}
\description{
Generate parameter map for phylogeny nll
}

