% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{gen_param_map_phylo_bd}
\alias{gen_param_map_phylo_bd}
\title{Generate parameter map for joint phylogeny and birth-death nll}
\usage{
gen_param_map_phylo_bd(tree, tip_times)
}
\arguments{
\item{tree}{tree with branch lengths in time units}

\item{tip_times}{times of tips of tree}
}
\description{
Generate parameter map for joint phylogeny and birth-death nll
}

