% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{get_time_tree_internal_nodeheights}
\alias{get_time_tree_internal_nodeheights}
\title{Get heights of internal nodes consistent with the tipheight,
substitution rate, and edge lengths in substitutions per site.}
\usage{
get_time_tree_internal_nodeheights(subs_tree, subs_per_time, tipheights)
}
\arguments{
\item{subs_tree}{tree with branch lengths measured in expected
substitutions per site}

\item{subs_per_time}{expected number of substitutions per unit time}

\item{tipheights}{heights of tips in time}
}
\description{
Get heights of internal nodes consistent with the tipheight,
substitution rate, and edge lengths in substitutions per site.
}

