% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth-death.R
\name{sim_bd_proc}
\alias{sim_bd_proc}
\title{Simulate tree according to a multi-type birth-death process}
\usage{
sim_bd_proc(n, l, m, psi, init = 1)
}
\arguments{
\item{n}{number of tips in simulated tree}

\item{l}{matrix of birth rates of each type for each type in process}

\item{m}{vector of death rates for each type. These deaths are not sampled.}

\item{psi}{vector of death rates for each type. These deaths are
followed by sampling.}

\item{init}{type of individual that starts the population}
}
\description{
Simulate tree according to a multi-type birth-death process
}

