% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularization.R
\name{stabpath_gpnet}
\alias{stabpath_gpnet}
\title{Calculate the stability path for a gpnet model}
\usage{
stabpath_gpnet(y, penalty.factor, size = 0.632, steps = 100, weakness = 1,
  mc.cores = getOption("mc.cores", 2L), ...)
}
\description{
Calculate the stability path for a gpnet model
}

